replaced all size_t

+++ lyke_individual.cpp
@@ -10,14 +10,14 @@ std::bitset<L> Individual::isNonsynSite; //global bitstring
 void Individual::init()
 {
 	isNonsynSite.reset();		 //sets every bit to 0
-	for (size_t i = 0u; i < L; i+=2)//sets every two bits (even) to 1
+	for (int i = 0; i < L; i+=2)//sets every two bits (even) to 1
 		isNonsynSite.set(i);
 }
 
 Individual::Individual() //Object Individual() of class Individual
   : x{}, y{}, z{}, ecotype{0.0}
 {
-	for (size_t i = 0u; i < L; ++i)
+	for (int i = 0; i < L; ++i)
 	{
 		if(rnd::uniform() < 0.5) x.set(i);
 		//uniform distribution with a likelihood of 0.5 that bits of x are set to 1.
@@ -37,7 +37,7 @@ Individual::Individual(Individual const * const mother, Individual const * const
 	// likelihood of 0.5 to have the x from the mother/father Individual
 	y = rnd::uniform() < 0.5 ? mother->y : father->y;
 	// likelihood of 0.5 to have the y from the mother/father Individual
-	for (size_t i = 0u; i < nGeneEco; ++i)
+	for (int i = 0; i < nGeneEco; ++i)
 		z.push_back(rnd::uniform() < 0.5 ? mother->z[i] : father->z[i]);
 	//Generates a mix of the "ecological genes", from two Individuals
 	mutate(); // Flips a bit in the bitstrings + the ecological character z
@@ -74,7 +74,7 @@ void Individual::develop()
 //The phenotype based on the genotype of the ecological character
 {
 	double sum = 0.0;
-	for (size_t i = 0u; i < nGeneEco; ++i)
+	for (int i = 0; i < nGeneEco; ++i)
 		sum += z[i];
 	ecotype = sum / nGeneEco;
 }
@@ -109,7 +109,7 @@ void Individual::print() const //output
 	std::cout << "y = " << ",";
 	std::cout << y << '\n';
 	std::cout << "z = " << ",";
-	for (size_t i = 0u; i < nGeneEco; ++i) std::cout << z[i] << ' ';
+	for (int i = 0; i < nGeneEco; ++i) std::cout << z[i] << ' ';
 	std::cout << '\n';
 	std::cout << "ecotype = " << ",";
 	std::cout << ecotype << '\n' << std::endl;
